Explicitly cast from Number to Double in Utils.getSAMQuality()
In v0.5.0, PathObject.getMeasurements() returns Number instead of Double.

+++ src/main/java/org/elephant/sam/Utils.java
@@ -183,7 +183,8 @@ public class Utils {
      * @return the quality score, or null if no score could be found
      */
     static Double getSAMQuality(PathObject pathObject) {
-        return pathObject.getMeasurements().getOrDefault(SAM_QUALITY_MEASUREMENT, null);
+        return (Double) pathObject.getMeasurements()
+                .getOrDefault(SAM_QUALITY_MEASUREMENT, null);
     }
 
     /**
